Test java8 stream
java8 stream API is lazy evaluator, which is similar with LINQ in c#.
this test is test when the stream will be evaluated.

When use kafka stream, i was curious that the stream code looks like
immediate execute, but the stream data is not yet  produce. so searched
lots of key word”kafka stream work” in google, finally, it is the
feature in java8.